Move test from test_distributions.py to test_distributions_util.py
