+++ test/test_distributions_util.py
@@ -12,6 +12,11 @@ from jax import lax, random, vmap
 import jax.numpy as jnp
 from jax.scipy.special import expit, xlog1py, xlogy
 
+import jax
+import numpyro
+import numpyro.distributions as dist
+from numpyro.infer import MCMC, NUTS, Predictive
+
 from numpyro.distributions.util import (
     add_diag,
     binary_cross_entropy_with_logits,
@@ -23,6 +28,8 @@ from numpyro.distributions.util import (
     vec_to_tril_matrix,
     von_mises_centered,
 )
+
+numpyro.set_host_device_count(2)
 
 
 @pytest.mark.parametrize("x, y", [(0.2, 10.0), (0.6, -10.0)])
@@ -182,3 +189,42 @@ def test_add_diag(matrix_shape: tuple, diag_shape: tuple) -> None:
     expected = matrix + diag[..., None] * jnp.eye(matrix.shape[-1])
     actual = add_diag(matrix, diag)
     np.testing.assert_allclose(actual, expected)
+
+
+@pytest.mark.parametrize(
+    "my_dist",
+    [
+        dist.TruncatedNormal(low=-1, high=1),
+        dist.TruncatedCauchy(low=-1, high=1),
+    ],
+)
+def test_no_tracer_leakage_in_truncated_distribution(my_dist):
+    """
+    Tests parallel sampling and use of multiple predictve methods
+    on models using truncated distributions.
+    Reference: https://github.com/pyro-ppl/numpyro/issues/1836, and
+    https://github.com/CDCgov/multisignal-epi-inference/issues/282
+    """
+    n_data_samples = 10
+    n_prior_samples = 10
+    n_mcmc_samples = 10
+    n_mcmc_warmup = 10
+    n_mcmc_chains = 2
+    rng_key = jax.random.PRNGKey(0)
+
+    data_samples = my_dist.sample(rng_key, (n_data_samples,))
+
+    def my_model(obs=None):
+        numpyro.sample("obs", my_dist, obs=obs)
+
+    prior_predictive = Predictive(my_model, num_samples=n_prior_samples)
+    prior_predictive(rng_key)
+
+    nuts_kernel = NUTS(my_model)
+    mcmc = MCMC(
+        nuts_kernel,
+        num_warmup=n_mcmc_warmup,
+        num_samples=n_mcmc_samples,
+        num_chains=n_mcmc_chains,
+    )
+    mcmc.run(rng_key, obs=data_samples)
